@@ -2,9 +2,10 @@
   "name": "chartjs-chart-financial",
   "description": "Chart.js module for charting financial securities",
   "main": "dist/chartjs-chart-financial.min.js",
+  "version": "0.2.0",
+  "type": "module",
   "module": "dist/chartjs-chart-financial.esm.js",
   "types": "types/index.d.ts",
-  "version": "0.1.0",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -15,29 +16,31 @@
     "types/*.d.ts"
   ],
   "scripts": {
-    "build": "rollup -c && cp dist/chartjs-chart-financial.js docs/",
+    "build": "rollup -c",
     "package": "npm run build && node scripts/create-packages.js",
     "lint": "eslint docs/**/*.html src/**/*.js test/**/*.js",
     "lint-types": "eslint \"types/**/*.ts\" && tsc -p types/tests/",
     "test": "npm run test-types && npm run test-unit",
-    "test-unit": "karma start --single-run --coverage --grep",
-    "test-unit:dev": "karma start --auto-watch --grep",
+    "test-unit": "karma start ./karma.conf.cjs --single-run --coverage --grep",
+    "test-unit:dev": "karma start ./karma.conf.cjs --auto-watch --grep",
     "test-types": "tsc -p types/tests/"
   },
   "peerDependencies": {
-    "chart.js": "^3.2.0 || ^4.0.0"
+    "chart.js": "^4.0.0"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.22.0",
-    "@typescript-eslint/parser": "^4.22.0",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@typescript-eslint/eslint-plugin": "^5.32.0",
+    "@typescript-eslint/parser": "^5.32.0",
     "archiver": "^5.3.0",
-    "chart.js": "^3.2.0",
-    "chartjs-adapter-luxon": "^1.0.0",
+    "chart.js": "^4.4.1",
+    "chartjs-adapter-luxon": "^1.3.1",
     "cross-env": "^7.0.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.56.0",
     "eslint-config-chartjs": "^0.3.0",
     "eslint-plugin-es": "^4.1.0",
-    "eslint-plugin-html": "^6.1.2",
+    "eslint-plugin-html": "^8.0.0",
+    "eslint-plugin-markdown": "^3.0.1",
     "jasmine": "^3.9.0",
     "karma": "^6.3.4",
     "karma-chrome-launcher": "^3.1.0",
@@ -46,10 +49,10 @@
     "karma-jasmine": "^4.0.1",
     "karma-jasmine-html-reporter": "^1.7.0",
     "karma-rollup-preprocessor": "^7.0.7",
-    "luxon": "^2.0.2",
+    "luxon": "^3.4.4",
     "rollup": "^2.56.3",
     "rollup-plugin-istanbul": "^3.0.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "typescript": "^4.2.4"
+    "typescript": "^4.2.4",
+    "yargs": "^17.7.2"
   }
 }
