@@ -1,124 +1,114 @@
 var barCount = 60;
-var initialDateStr = '01 Apr 2017 00:00 Z';
+var initialDateStr = new Date().toUTCString();
 
 var ctx = document.getElementById('chart').getContext('2d');
 ctx.canvas.width = 1000;
 ctx.canvas.height = 250;
 
-var barData = getRandomData(initialDateStr, barCount);
-function lineData() { return barData.map(d => { return { x: d.x, y: d.c} }) };
+var barData = new Array(barCount);
+var lineData = new Array(barCount);
+
+getRandomData(initialDateStr);
 
 var chart = new Chart(ctx, {
-	type: 'candlestick',
-	data: {
-		datasets: [{
-			label: 'CHRT - Chart.js Corporation',
-			data: barData
-		}]
-	}
+  type: 'candlestick',
+  data: {
+    datasets: [{
+      label: 'CHRT - Chart.js Corporation',
+      data: barData,
+    }, {
+      label: 'Close price',
+      type: 'line',
+      data: lineData,
+      hidden: true,
+    }]
+  }
 });
 
-var getRandomInt = function(max) {
-	return Math.floor(Math.random() * Math.floor(max));
-};
-
 function randomNumber(min, max) {
-	return Math.random() * (max - min) + min;
+  return Math.random() * (max - min) + min;
 }
 
-function randomBar(date, lastClose) {
-	var open = +randomNumber(lastClose * 0.95, lastClose * 1.05).toFixed(2);
-	var close = +randomNumber(open * 0.95, open * 1.05).toFixed(2);
-	var high = +randomNumber(Math.max(open, close), Math.max(open, close) * 1.1).toFixed(2);
-	var low = +randomNumber(Math.min(open, close) * 0.9, Math.min(open, close)).toFixed(2);
-	return {
-		x: date.valueOf(),
-		o: open,
-		h: high,
-		l: low,
-		c: close
-	};
+function randomBar(target, index, date, lastClose) {
+  var open = +randomNumber(lastClose * 0.95, lastClose * 1.05).toFixed(2);
+  var close = +randomNumber(open * 0.95, open * 1.05).toFixed(2);
+  var high = +randomNumber(Math.max(open, close), Math.max(open, close) * 1.1).toFixed(2);
+  var low = +randomNumber(Math.min(open, close) * 0.9, Math.min(open, close)).toFixed(2);
+
+  if (!target[index]) {
+    target[index] = {};
+  }
+
+  Object.assign(target[index], {
+    x: date.valueOf(),
+    o: open,
+    h: high,
+    l: low,
+    c: close
+  });
 
 }
 
-function getRandomData(dateStr, count) {
-	var date = luxon.DateTime.fromRFC2822(dateStr);
-	var data = [randomBar(date, 30)];
-	while (data.length < count) {
-		date = date.plus({days: 1});
-		if (date.weekday <= 5) {
-			data.push(randomBar(date, data[data.length - 1].c));
-		}
-	}
-	return data;
+function getRandomData(dateStr) {
+  var date = luxon.DateTime.fromRFC2822(dateStr);
+
+  for (let i = 0; i < barData.length;) {
+    date = date.plus({days: 1});
+    if (date.weekday <= 5) {
+      randomBar(barData, i, date, i === 0 ? 30 : barData[i - 1].c);
+      lineData[i] = {x: barData[i].x, y: barData[i].c};
+      i++;
+    }
+  }
 }
 
 var update = function() {
-	var dataset = chart.config.data.datasets[0];
+  var dataset = chart.config.data.datasets[0];
 
-	// candlestick vs ohlc
-	var type = document.getElementById('type').value;
-	dataset.type = type;
+  // candlestick vs ohlc
+  var type = document.getElementById('type').value;
+  chart.config.type = type;
 
-	// linear vs log
-	var scaleType = document.getElementById('scale-type').value;
-	chart.config.options.scales.y.type = scaleType;
+  // linear vs log
+  var scaleType = document.getElementById('scale-type').value;
+  chart.config.options.scales.y.type = scaleType;
 
-	// color
-	var colorScheme = document.getElementById('color-scheme').value;
-	if (colorScheme === 'neon') {
-		dataset.color = {
-			up: '#01ff01',
-			down: '#fe0000',
-			unchanged: '#999',
-		};
-	} else {
-		delete dataset.color;
-	}
+  // color
+  var colorScheme = document.getElementById('color-scheme').value;
+  if (colorScheme === 'neon') {
+    chart.config.data.datasets[0].backgroundColors = {
+      up: '#01ff01',
+      down: '#fe0000',
+      unchanged: '#999',
+    };
+  } else {
+    delete chart.config.data.datasets[0].backgroundColors;
+  }
 
-	// border
-	var border = document.getElementById('border').value;
-	var defaultOpts = Chart.defaults.elements[type];
-	if (border === 'true') {
-		dataset.borderColor = defaultOpts.borderColor;
-	} else {
-		dataset.borderColor = {
-			up: defaultOpts.color.up,
-			down: defaultOpts.color.down,
-			unchanged: defaultOpts.color.up
-		};
-	}
+  // border
+  var border = document.getElementById('border').value;
+  if (border === 'false') {
+    dataset.borderColors = 'rgba(0, 0, 0, 0)';
+  } else {
+    delete dataset.borderColors;
+  }
 
-	// mixed charts
-	var mixed = document.getElementById('mixed').value;
-	if(mixed === 'true') {
-		chart.config.data.datasets = [
-			{
-				label: 'CHRT - Chart.js Corporation',
-				data: barData
-			},
-			{
-				label: 'Close price',
-				type: 'line',
-				data: lineData()
-			}	
-		]
-	}
-	else {
-		chart.config.data.datasets = [
-			{
-				label: 'CHRT - Chart.js Corporation',
-				data: barData
-			}	
-		]
-	}
+  // mixed charts
+  var mixed = document.getElementById('mixed').value;
+  if (mixed === 'true') {
+    chart.config.data.datasets[1].hidden = false;
+  } else {
+    chart.config.data.datasets[1].hidden = true;
+  }
 
-	chart.update();
+  chart.update();
 };
+
+[...document.getElementsByTagName('select')].forEach(element => element.addEventListener('change', update));
 
 document.getElementById('update').addEventListener('click', update);
 
 document.getElementById('randomizeData').addEventListener('click', function() {
-	barData = getRandomData(initialDateStr, barCount);
-	update();
+  getRandomData(initialDateStr, barData);
+  update();
 });
