@@ -1,47 +1,33 @@
-﻿'use strict';
-
-import {Chart} from 'chart.js';
-import {merge} from 'chart.js/helpers';
-import {FinancialController} from './controller.financial';
+﻿import {FinancialController} from './controller.financial';
 import {CandlestickElement} from './element.candlestick';
+import {BarController} from 'chart.js';
 
 export class CandlestickController extends FinancialController {
 
-	updateElements(elements, start, count, mode) {
-		const me = this;
-		const dataset = me.getDataset();
-		const ruler = me._ruler || me._getRuler();
-		const firstOpts = me.resolveDataElementOptions(start, mode);
-		const sharedOptions = me.getSharedOptions(firstOpts);
-		const includeOptions = me.includeOptions(mode, sharedOptions);
+  static id = 'candlestick';
 
-		me.updateSharedOptions(sharedOptions, mode, firstOpts);
+  static defaults = {
+    ...FinancialController.defaults,
+    dataElementType: CandlestickElement.id
+  };
 
-		for (let i = start; i < count; i++) {
-			const options = sharedOptions || me.resolveDataElementOptions(i, mode);
+  static defaultRoutes = BarController.defaultRoutes;
 
-			const baseProperties = me.calculateElementProperties(i, ruler, mode === 'reset', options);
-			const properties = {
-				...baseProperties,
-				datasetLabel: dataset.label || '',
-				// label: '', // to get label value please use dataset.data[index].label
+  updateElements(elements, start, count, mode) {
+    const reset = mode === 'reset';
+    const ruler = this._getRuler();
+    const {sharedOptions, includeOptions} = this._getSharedOptions(start, mode);
 
-				// Appearance
-				color: dataset.color,
-				borderColor: dataset.borderColor,
-				borderWidth: dataset.borderWidth,
-			};
+    for (let i = start; i < start + count; i++) {
+      const options = sharedOptions || this.resolveDataElementOptions(i, mode);
 
-			if (includeOptions) {
-				properties.options = options;
-			}
-			me.updateElement(elements[i], i, properties, mode);
-		}
-	}
+      const baseProperties = this.calculateElementProperties(i, ruler, reset, options);
+
+      if (includeOptions) {
+        baseProperties.options = options;
+      }
+      this.updateElement(elements[i], i, baseProperties, mode);
+    }
+  }
 
 }
-
-CandlestickController.id = 'candlestick';
-CandlestickController.defaults = merge({
-	dataElementType: CandlestickElement.id
-}, Chart.defaults.financial);
